ntdll/signalx86_64: Add CFI for __wine_syscall_dispatcher
In a debugger that is capable of loading both PE and ELF libraries
(e.g. the wine-enhanced GDB version at [1]), this unwind information
will allow proper unwind across the __wine_syscall_dispatcher.

The CFI itself is relatively straightforward, though the implementation
looks a little complex, because support for proper CFI encoding is
somewhat limitied in the GNU toolchain. Essentially, as registers are
stored in the rcx buffer in this function, we switch their cfi
description to point at that rcx buffer, so the registers may be
freely used thereafter without affecting the unwind. Eventually the
rcx buffer is tracked in rbp instead at which point we pivot the
CFI information as well and keep it there for most of the rest of
the function.

[1] https://github.com/JuliaComputing/gdb-solib-wine

Signed-off-by: Keno Fischer <keno@juliacomputing.com>